// frontend/lib/client/models/page.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Clipboard API
 * REST API for managing clipboard entries
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { PageableObject } from './pageable-object';
// May contain unused imports in some cases
// @ts-ignore
import { SortObject } from './sort-object';

/**
 * 
 * @export
 * @interface Page
 */
export interface Page {
    /**
     * 
     * @type {number}
     * @memberof Page
     */
    'totalElements'?: number;
    /**
     * 
     * @type {number}
     * @memberof Page
     */
    'totalPages'?: number;
    /**
     * 
     * @type {number}
     * @memberof Page
     */
    'size'?: number;
    /**
     * 
     * @type {Array<any>}
     * @memberof Page
     */
    'content'?: Array<any>;
    /**
     * 
     * @type {number}
     * @memberof Page
     */
    'number'?: number;
    /**
     * 
     * @type {number}
     * @memberof Page
     */
    'numberOfElements'?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof Page
     */
    'pageable'?: PageableObject;
    /**
     * 
     * @type {SortObject}
     * @memberof Page
     */
    'sort'?: SortObject;
    /**
     * 
     * @type {boolean}
     * @memberof Page
     */
    'first'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Page
     */
    'last'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Page
     */
    'empty'?: boolean;
}